requirements.txt
certifi==2022.9.24
charset-normalizer==3.1.0
cycler==0.11.0
easydict==1.10
einops==0.6.1
filelock==3.15.4
fonttools==4.39.4
fsspec==2024.6.1
huggingface-hub==0.23.4
idna==3.4
imageio==2.31.1
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.5.2
networkx==2.8.8
numpy==1.24.4
opencv-python==4.6.0.66
packaging==23.1
Pillow==10.0.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.31.0
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tensorboardX==2.6
thop 
tifffile==2023.7.10
timm==0.6.7
torch 
torchvision 
tqdm==4.66.1
typing_extensions==4.6.2
unzip==1.0.0
urllib3==2.0.2
